externals/cleandict/requirements.txt
opencc-python-reimplemented==0.1.1
google-generativeai==0.5.0
python-dotenv==1.0.0
